requirements.txt
pyTelegramBotAPI
psycopg2-binary
openai